+++ src/io/import_nirs.jl
@@ -101,10 +101,18 @@ function import_nirs(file_name::String)
     # stimuli
     s = nirs["s"][:]
     if s == zeros(length(time_pts))    
-        markers = DataFrame(:id=>String[], :start=>Int64[], :length=>Int64[], :description=>String[], :channel=>Int64[])
+        markers = DataFrame(:id=>String[],
+                            :start=>Float64[],
+                            :length=>Float64[],
+                            :description=>String[],
+                            :channel=>Int64[])
     else
         s_n = size(s, 2)
-        markers = DataFrame(:id=>String[], :start=>Int64[], :length=>Int64[], :description=>String[], :channel=>Int64[])
+        markers = DataFrame(:id=>String[],
+                            :start=>Float64[],
+                            :length=>Float64[],
+                            :description=>String[],
+                            :channel=>Int64[])
         for idx1 in 1:s_n
             s_start = findall(s[:, idx1] .!= 0.0)
             for idx2 in eachindex(s_start)
